backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
python-multipart
requests
python-dotenv
pymongo
motor
pandas
yfinance
pandas-ta
scikit-learn
statsmodels
prophet
backtesting
numpy
scipy
plotly
beautifulsoup4
lxml
ta
ccxt